feat: add V4 multipart proxy support and enhance analytics routing
- Add proxyMultipart function for file upload proxying with duplex streaming
- Update /contracts/bulk to use proxyMultipart instead of proxyJson
- Add /ingestion/bulk-upload endpoint with multipart support
- Add forwardedHeaders extraction for auth and trace ID correlation
- Enhance proxyJson to handle non-JSON responses and empty bodies gracefully
- Add v4EventStreams constants and export SSE/streams modules
- Add v4Analytics router

+++ backend/src/wcp_backend/api/analytics.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from fastapi import APIRouter, HTTPException, Query
 from pydantic import BaseModel
-from sqlalchemy import desc, func, select
+from sqlalchemy import desc, func, select, Integer
 
 from wcp_backend.config import settings
 from wcp_backend.services.db import async_session
@@ -178,7 +178,7 @@ async def approval_by_trade() -> dict[str, Any]:
                         func.case(
                             (decisions_table.c.verdict == "approved", 1),
                             else_=0,
-                        )
+                        ).cast(Integer)
                     ).label("approved_count"),
                 )
                 .group_by(decisions_table.c.trust_band)
